Windows comparison might be on a good way, but this path vs. filename thing is not working yet.

+++ win32_error_exception.h
@@ -72,7 +72,7 @@ private:
 			return;
 		}
 
-		msg = new char[utf8_size + 1];
+		msg = new char[(size_t)utf8_size + 1];
 		msg[utf8_size] = '\0';
 
 		utf8_size = WideCharToMultiByte(
